Transfer API code to api app and remove unnecessary HTTP functions
